helloworld/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3

google-auth
google-api-core
google-cloud-storage